split up builders and tests into seperate files and other review comments
