+++ src/op/builder/torch/lstm.cpp
@@ -1,4 +1,5 @@
-/* The MIT License (MIT)
+/*
+ * The MIT License (MIT)
  *
  * Copyright (c) 2015-2026 Advanced Micro Devices, Inc. All rights reserved.
  *
@@ -21,44 +22,59 @@
  * THE SOFTWARE.
  */
 
-#include <migraphx/common.hpp>
+#include <cstddef>
+#include <vector>
 #include <migraphx/instruction.hpp>
 #include <migraphx/make_op.hpp>
-#include <migraphx/tune_axis.hpp>
+#include <migraphx/operation.hpp>
+#include <migraphx/serialize.hpp>
 #include <migraphx/op/builder/op_builder.hpp>
-#include <migraphx/op/builder/insert.hpp>
+#include <migraphx/op/common.hpp>
 
 namespace migraphx {
 inline namespace MIGRAPHX_INLINE_NS {
 namespace op {
 namespace builder {
 
-// glu has no native op: split in half along `axis`, gate the first half by sigmoid(second).
-struct glu : op_builder<glu>
+// lstm expands into an lstm op plus its last hidden-state and last cell-state outputs.
+struct torch_lstm : op_builder<torch_lstm>
 {
-    int64_t axis = -1;
+    std::size_t hidden_size = 1;
+    std::vector<operation> actv_funcs{};
+    rnn_direction direction = rnn_direction::forward;
+    float clip              = 0.0f;
+    int input_forget        = 0;
 
     template <class Self, class F>
     static auto reflect(Self& self, F f)
     {
-        return pack(f(self.axis, "axis"));
+        return pack(f(self.hidden_size, "hidden_size"),
+                    f(self.actv_funcs, "actv_func"),
+                    f(self.direction, "direction"),
+                    f(self.clip, "clip"),
+                    f(self.input_forget, "input_forget"));
     }
+
+    static std::vector<std::string> names() { return {"tm::lstm"}; }
 
     std::vector<instruction_ref>
     insert(module& m, instruction_ref ins, const std::vector<instruction_ref>& args) const
     {
-        auto x    = args[0];
-        auto lens = x->get_shape().lens();
-        auto ax   = tune_axis(lens.size(), axis, "glu");
-        auto len  = static_cast<int64_t>(lens[ax]);
-        auto half = len / 2;
-
-        auto first = m.insert_instruction(
-            ins, make_op("slice", {{"axes", {ax}}, {"starts", {0}}, {"ends", {half}}}), x);
-        auto second = m.insert_instruction(
-            ins, make_op("slice", {{"axes", {ax}}, {"starts", {half}}, {"ends", {len}}}), x);
-        auto gate = m.insert_instruction(ins, make_op("sigmoid"), second);
-        return {m.insert_instruction(ins, make_op("mul"), first, gate)};
+        auto self = *this;
+        if(self.actv_funcs.empty())
+        {
+            self.actv_funcs = {make_op("sigmoid"), make_op("tanh"), make_op("tanh")};
+            if(self.direction == rnn_direction::bidirectional)
+            {
+                self.actv_funcs.insert(self.actv_funcs.end(),
+                                       {make_op("sigmoid"), make_op("tanh"), make_op("tanh")});
+            }
+        }
+        auto hidden_states =
+            m.insert_instruction(ins, make_op("lstm", migraphx::to_value(self)), args);
+        auto last_hs   = m.insert_instruction(ins, make_op("rnn_last_hs_output"), hidden_states);
+        auto last_cell = m.insert_instruction(ins, make_op("rnn_last_cell_output"), hidden_states);
+        return {hidden_states, last_hs, last_cell};
     }
 };
