+++ src/op/builder/torch/slice_scatter.cpp
@@ -1,4 +1,5 @@
-/* The MIT License (MIT)
+/*
+ * The MIT License (MIT)
  *
  * Copyright (c) 2015-2026 Advanced Micro Devices, Inc. All rights reserved.
  *
@@ -21,44 +22,51 @@
  * THE SOFTWARE.
  */
 
-#include <migraphx/common.hpp>
+#include <cstddef>
+#include <cstdint>
+#include <vector>
 #include <migraphx/instruction.hpp>
+#include <migraphx/literal.hpp>
 #include <migraphx/make_op.hpp>
 #include <migraphx/tune_axis.hpp>
 #include <migraphx/op/builder/op_builder.hpp>
-#include <migraphx/op/builder/insert.hpp>
 
 namespace migraphx {
 inline namespace MIGRAPHX_INLINE_NS {
 namespace op {
 namespace builder {
 
-// glu has no native op: split in half along `axis`, gate the first half by sigmoid(second).
-struct glu : op_builder<glu>
+// slice_scatter has no native op: scatter src into the [start:end:step] slice along `dim`.
+struct torch_slice_scatter : op_builder<torch_slice_scatter>
 {
-    int64_t axis = -1;
+    int64_t dim   = 0;
+    int64_t start = 0;
+    int64_t end   = 0;
+    int64_t step  = 1;
 
     template <class Self, class F>
     static auto reflect(Self& self, F f)
     {
-        return pack(f(self.axis, "axis"));
+        return pack(
+            f(self.dim, "dim"), f(self.start, "start"), f(self.end, "end"), f(self.step, "step"));
     }
+
+    static std::vector<std::string> names() { return {"tm::slice_scatter"}; }
 
     std::vector<instruction_ref>
     insert(module& m, instruction_ref ins, const std::vector<instruction_ref>& args) const
     {
-        auto x    = args[0];
-        auto lens = x->get_shape().lens();
-        auto ax   = tune_axis(lens.size(), axis, "glu");
-        auto len  = static_cast<int64_t>(lens[ax]);
-        auto half = len / 2;
+        shape idx_shape{shape::int64_type, args[1]->get_shape().lens()};
+        auto axis = tune_axis(idx_shape.ndim(), dim, "slice_scatter");
+        std::vector<int64_t> data(idx_shape.elements());
+        for(std::size_t i = 0; i < data.size(); ++i)
+            data[i] = start + step * idx_shape.multi(i)[axis];
+        auto indices = m.add_literal(literal{idx_shape, data.begin(), data.end()});
 
-        auto first = m.insert_instruction(
-            ins, make_op("slice", {{"axes", {ax}}, {"starts", {0}}, {"ends", {half}}}), x);
-        auto second = m.insert_instruction(
-            ins, make_op("slice", {{"axes", {ax}}, {"starts", {half}}, {"ends", {len}}}), x);
-        auto gate = m.insert_instruction(ins, make_op("sigmoid"), second);
-        return {m.insert_instruction(ins, make_op("mul"), first, gate)};
+        auto std_input = m.insert_instruction(ins, make_op("contiguous"), args[0]);
+        auto std_src   = m.insert_instruction(ins, make_op("contiguous"), args[1]);
+        return {m.insert_instruction(
+            ins, make_op("scatter_none", {{"axis", axis}}), {std_input, indices, std_src})};
     }
 };
